Removed some errors when compiling with "-Werror".
git-svn-id: svn://projects.linuxtogo.org/svn/gpe/trunk@7707 d4735bf7-8e1f-0410-bce4-ef584eb11c25

+++ base/gpe-opensync/src/gpesync_client.c
@@ -125,7 +125,7 @@ read_lines (struct gpesync_client_query_context *query_ctx, char *data)
 
       /* We need to split the data in the sperate lines, so that
        * we can read it as lists. */
-      int count = 0;
+      gsize count = 0;
       line_iter = data;
       line = get_next_line (line_iter, &count);
       do
